Add status bar padding to list if there are not featured posts
Featured posts carry status bar padding to make sure we are drawing blured image and gradients behind status bar. As such when there are no featured posts and only list, the status bar padding is not applied. This ensures we apply the padding even when there are no featured posts