Add offset progress in bottom sheet scaffold
While progress + bottom sheet value does provide information on what's the state of bottom sheet expanding/collapsing. For animations it's not that smooth to observer. Since the progress only triggers after a specific threshold. So added a variable to listen to offset changes, so that animations in bottom sheet can use this instead