Use combined click for triggering onClick and onLongClick in feed list item
For some reason `pointerInput` is having issues. For example, if I have 10 items and 9 of them are on screen and once I scroll to the 10 item and try the long press the display dropdown, it's not working. The long press is triggered by dropdown is not shown, probably focus related issue? I will have to debug it later