Fix app crashing when sharing post/feed on an iPad
I tried placing the popover anchor to be based on the share icon root offset. But for some reason the offset from onGloballyPositioned is different from the bounds of the viewcontroller view. I am lazy to debug it and resolve it fully, so showing it in bottom center for the screen instead similar to how the sheet is shown on mobile.

fixes #96