Catch errors when trying to create image from encoded data on iOS
It's a quick fix for crash that happens when the image URL contains a gif or video content or any thing other than image. Ideally we would want to handle those or ignore them during parsing. But a quick way to handle them would be just ignore them during rendering.