Stop creating new feeds pager every time selected feed is changed in feeds sheet
This was causing an issue where the feeds list scroll position is getting reset. It's really annoying when it's in bottom bar mode, when user is switching between feeds.

Unlike posts pager created in home presenter this doesn't consume selected feed, so we don't have to create it inside the flat map