Move background workers on Android back to root package (#174)
Since we already have an existing work scheduled in existing installs, those will not have `background` package in the class name, which leads to a `IllegalArgumentException`. I can cancel existing workers and have a proper migration for it. But it seems to much effort for a little gain. So moving them back to the root package for now.