Stop removing `www.` when transforming the URL for fetching the feed (#175)
I assumed all website will redirect to URL with `www.` if we pass a URL without one. But that doesn't seem to be the case. It looks like some websites fail with SSL handshake error. So, we are keeping the `www.`. There is an off chance, where some domains are not properly configured to open with `www.` (I doubt there are many), but I can handle it if I see those errors.