Lock clock when testing date time formatters
This is done to fix the failing date formatters test. The reason it fails is, since one of the date format we have doesn't have year attached to it "MM-dd HH:mm:ss", it picks it from the current system time. 

The expected value I provided was for year 2023, but since the year is now 2024, the tests has started to fail. So locking the clock (which I should've done from the start) to a specific date time, so that the tests are reproducible.