Open reader view when a post is clicked (#233)
* Add webview dependency

* Add query to get post based on link

* Add readability dependency to shared module

* Add initial implementation of reader screen

* Navigate to reader screen when post is clicked in home screen

* Remove unused code from `HomePresenter`

* Open reader screen when post is clicked in search screen

* Open reader screen when post is clicked in bookmarks screen

* Remove unused code

* Add a TODO in reader screen

* Add JS bridge to open links from reader view

* Add padding to `WebView` composable to handle navigation gestures

Since webview is consuming the navigation gestures on iOS when it's edge to edge. Adding padding to the composable so that the system gestures are still handled properly

* Change `RssRepository#feed` function to a suspend function

* Load feed info when reader screen is initialised

* Add feed name and link to the reader view layout

* Clean up reader view HTML styles

* Change feed link to feed home page link in the reader view

* Add icons in bottom bar in reader view

* Add support for book marking and un-bookmarking post in reader view

* Handle nested clickable elements in reader view

In certain scenarios the link text is formatted and the text element doesn't contain a HREF, so we look for it and then open it.

* Hard code kotlin-test version

For some weird reason after adding the webview dependency kotlin-test dependency sync is having issue and this is the only thing that fixed. Will look into it further later.