Call `HomeScreen.Init` when `HomePresenterInstanse` class init happens (#273)
I don't know why I decided to do it in on lifecycle on create. That ended up calling this `init` every time home screen presenter is created. Which means the paging cache is not used. Instead we are using the presenter instance class init to trigger subscribing to cached paging flow which fixes the scroll position getting reset of config changes. Weirdly it doesn't fix scroll position getting reset for navigation. Will have to check separately and resolve it in a different PR.