Consume crashes in `HtmlContentParser#parse` (#322)
* Add CrashKiOS bugsnag dependency to network module

* Set BugSnag custom value when fetching feed

This would make sure we receive the feed URL associated with the crash irrespective of where it's called from. That way we don't have to set this in multiple places like we do now.

* Remove sending feed URL as custom value to BugSnag in the home presenter

* Wrap entire `HtmlContentParser` function in try/catch

Rather than crashing the app if parsing the HTML fails at any point, we just return null.

We are also reporting the MalformedInputException now to Bugsnag to get more info about the types of feeds that cause this. So, that I can handle those sort of feeds better