Update HomePresenter and UI for current date handling (#983)
* Refactor date extensions

Might have to move this to a separate helper class at some point.

* Update current date time on home screen resume

We are also using the tracked current date time when fetching posts now

* Refresh posts list in the home screen when current date time is changed

* Display current date in the home app bar title

When no source is selected, we display the current date in the app bar

* Update HomePresenter state when date changes

The `updateCurrentDateTime` function in `HomePresenter` now only updates the state if the new date is greater than the current date in the state.