Fix reader screen opening incorrect post (#1037)
* Pass initial post id to the reader screen

Instead of post index, we will use post id to determine the page index
since pagination doesn't load entire list of posts and loads subset of
it.

* Calculate adjusted initial page in the reader screen

Since pagination only loads subset of posts in a page, post index we
received from the home screen is not directly mapped to the posts page
we have in the reader screen.

So we are using item snapshot list to calculate the adjusted initial
page to show. Maybe calculating this with `rowId` in a SQL query might
be better? But at this point this works and we don't need another query
to calculate the initial page.

We are still using the post index we received to let the `Pager` know
what's the initial key/index we are trying to load from.