Migrate iOS app database to app group
This change moves the application's database file to an app group container on iOS. This is necessary to allow other app extensions, like widgets, to access the same database.

The migration logic checks for the existence of the database in the old location (application support directory) and moves it to the new app group container path if found. It also moves the associated WAL (Write-Ahead Log) and SHM (Shared Memory) files.

The database path is now configurable, pointing to the app group container. If the app group database path doesn't exist, it will be created.