Refactor feed content parsers (#1243)
* Refactor HtmlContentParser to be a class

The HtmlContentParser was previously an object, which made it difficult to inject dependencies if needed in the future. This commit refactors it into a class and updates its usages.

* Rename HtmlContentParser to ArticleHtmlParser

This commit renames `HtmlContentParser` to `ArticleHtmlParser` to better reflect its purpose of parsing HTML content specifically for articles. The functionality of the class remains the same.

* Rename feed content parsers

Renamed `RSSContentParser` to `RssFeedParser`, `RDFContentParser` to `RDFFeedParser`, and `AtomContentParser` to `AtomFeedParser` for clarity and consistency.