Move post raw content to a separate table
The `rawContent` field has been removed from the `Post` table and moved to a new `postContent` table. This change helps to keep the main `Post` table lean, as `rawContent` can be large and is not always needed.

The `postContent` table will store the `rawContent` and `htmlContent` (for full article view) for each post.

This change also involves:
- Updating database queries and triggers to reflect the new table structure.
- Modifying `RssRepository` to handle the new `postContent` table.
- Creating a new `PostContentRepository` to manage data in the `postContent` table.
- Updating data models (`Post`, `PostWithMetadata`) to remove `rawContent`.
- Creating a new `PostContent` data model.