Move post content and full article HTML to separate table (#1281)
* Add `FullArticleFetcher` to fetch HTML content

* Move post raw content to a separate table

The `rawContent` field has been removed from the `Post` table and moved to a new `postContent` table. This change helps to keep the main `Post` table lean, as `rawContent` can be large and is not always needed.

The `postContent` table will store the `rawContent` and `htmlContent` (for full article view) for each post.

This change also involves:
- Updating database queries and triggers to reflect the new table structure.
- Modifying `RssRepository` to handle the new `postContent` table.
- Creating a new `PostContentRepository` to manage data in the `postContent` table.
- Updating data models (`Post`, `PostWithMetadata`) to remove `rawContent`.
- Creating a new `PostContent` data model.

* Refactor Reader JS to not fetch full article content

The functionality to fetch the full article content was removed from the `parseReaderContent` JavaScript function. Instead, the HTML content is now passed directly as a parameter.

This change simplifies the JavaScript logic and ensures that the reader component relies on the caller to provide the complete HTML content. Corresponding updates were made in the Android and iOS `ReaderWebView` implementations to pass the content directly.

* Run code formatting

* Remove image from formatted Reddit post

The image was previously displayed above the post content. This change removes the image from the formatted Reddit post since we already display the hero image in the detail view this always gets removed in later step, so removing it when formatting the post itself.

* Remove `fetchFullArticle` parameter from `ReaderWebView`

The `fetchFullArticle` parameter was unused and has been removed from the `ReaderWebView` composable function and its platform-specific implementations.

* Add vertical padding to progress indicator in reader screen

* Pass post ID to `ReaderPageViewModel`

The `ReaderPageViewModel` now takes the post ID as a constructor parameter. This allows the view model to load the post content and full article HTML directly, instead of relying on the `ReaderScreen` to pass it down.

* Load full article content in reader screen when the button is clicked

* Refactor `ReaderPageViewModel`