Add additional timestamps columns to the `Post` table
This change is done to allow us to track

`createdAt` - when post is first inserted in the table.
`updatedAt` - when post is last updated at. ATM this is same as
`createdAt`, but we can change the upsert behaviour to track post
updates.
`syncedAt` - this will track when the last sync happend when using 3rd
party integrations like FreshRSS. ATM this is same as `updatedAt`.

Renamed `date` to `postDate`, to better represent the date on which the
post is published.