Rename PostMetadata to PostActionBar
The `PostMetadata` composable has been renamed to `PostActionBar` to better reflect its purpose, as it contains action buttons for a post.

Following this, `PostOptionsButtonRow` has been renamed to `PostActions`. The `postPublishedAt` parameter is now `postRelativeTimestamp` for improved clarity. Additionally, the timestamp text is now uppercase and uses the `labelSmall` text style.