Use app dynamic color state in reader screen
Use the `LocalDynamicColorState` to provide default light and dark color schemes in the reader instead of `LocalAppColorScheme`.
This ensures that the correct theme is applied before the dynamic content theme is generated.

Additionally, this change prevents fetching post images for generating dynamic colors when the "Block images" setting is enabled.