Add setting to block downloading or showing images in the app (#1402)
* Add option to block images in settings

* Switch home view mode to simple when toggling image blocking

Automatically update the home view mode to simple whenever the block images setting is changed. This ensures the user interface remains consistent and optimized for a text-only experience when images are disabled.

* Don't load images when "Block images" setting is enabled

This will prevent images from being loaded and displayed in the reader view and feed list when the user has enabled the "Block images" option in the settings.

* Set HomeViewMode to Simple only when it's Default

When toggling the "Block images" setting, the `HomeViewMode` will now only be changed to `Simple` if the current view mode is `Default`.

* Use Coil AsyncImage directly for displaying about images

* Delete unnecessary ReaderImage

* Support blocking images in reader view

Convert `CoilMarkdownTransformer` to a class to handle `shouldBlockImage` state and suppress image rendering when enabled. Update `ReaderPage` to initialize the transformer with the current block image setting.

* Use app dynamic color state in reader screen

Use the `LocalDynamicColorState` to provide default light and dark color schemes in the reader instead of `LocalAppColorScheme`.
This ensures that the correct theme is applied before the dynamic content theme is generated.

Additionally, this change prevents fetching post images for generating dynamic colors when the "Block images" setting is enabled.