Add `useAmoled` option to dynamic colors
This commit introduces a `useAmoled` flag to the `DynamicColorState`.

When `useAmoled` is enabled, the generated dark theme color schemes will use a pure black background. This option is also incorporated into the color scheme caching key to ensure correct themes are stored and retrieved.

Additionally, a `refresh` method has been added to `DynamicColorState` to allow for the re-application of the last animated color transition, which is useful for updating the theme when the `useAmoled` setting changes. The `reset` method is also updated to correctly apply the amoled theme if enabled.